+++ pennylane_lightning/lightning_qubit/lightning_qubit.py
@@ -75,6 +75,11 @@ if LQ_CPP_BINARY_AVAILABLE:
         VectorJacobianProductC128,
     )
 
+    def _state_dtype(dtype):
+        if dtype not in [np.complex128, np.complex64]:  # pragma: no cover
+            raise ValueError(f"Data type is not supported for state-vector computation: {dtype}")
+        return StateVectorC128 if dtype == np.complex128 else StateVectorC64
+
     allowed_operations = {
         "Identity",
         "BasisState",
@@ -226,10 +231,7 @@ if LQ_CPP_BINARY_AVAILABLE:
 
             # Create the initial state. Internally, we store the
             # state as an array of dimension [2]*wires.
-            self._state = self._create_basis_state(0)
-            self._pre_rotated_state = self._state
-            self._c_dtype = c_dtype
-
+            self._qubit_state = _state_dtype(c_dtype)(self.num_wires)
             self._batch_obs = batch_obs
             self._mcmc = mcmc
             if self._mcmc:
@@ -274,24 +276,16 @@ if LQ_CPP_BINARY_AVAILABLE:
         def _create_basis_state(self, index):
             """Return a computational basis state over all wires.
             Args:
-                index (int): integer representing the computational basis state
-            Returns:
-                array[complex]: complex array of shape ``[2]*self.num_wires``
-                representing the statevector of the basis state
-            Note: This function does not support broadcasted inputs yet.
+                index (int): integer representing the computational basis state.
             """
-            state = allocate_aligned_array(2**self.num_wires, np.dtype(self.C_DTYPE), True)
-            state[index] = 1
-            return self._reshape(state, [2] * self.num_wires)
+            self._qubit_state.setBasisState(index)
 
         def reset(self):
             """Reset the device"""
             super().reset()
 
             # init the state vector to |00..0>
-            if not self.state[0] == 1.0 + 0j:
-                self._state = self._create_basis_state(0)
-            self._pre_rotated_state = self._state
+            self._qubit_state.resetStateVector()
 
         @property
         def create_ops_list(self):
@@ -301,25 +295,32 @@ if LQ_CPP_BINARY_AVAILABLE:
         @property
         def measurements(self):
             """Returns a Measurements object matching ``use_csingle`` precision."""
-            ket = np.ravel(self._state)
-            state_vector = StateVectorC64(ket) if self.use_csingle else StateVectorC128(ket)
             return (
-                MeasurementsC64(state_vector)
+                MeasurementsC64(self.state_vector)
                 if self.use_csingle
-                else MeasurementsC128(state_vector)
+                else MeasurementsC128(self.state_vector)
             )
 
         @property
         def state(self):
-            """Returns the flattened state vector."""
-            shape = (1 << self.num_wires,)
-            return self._reshape(self._pre_rotated_state, shape)
+            """Copy the state vector data to a numpy array.
+
+            **Example**
+
+            >>> dev = qml.device('lightning.kokkos', wires=1)
+            >>> dev.apply([qml.PauliX(wires=[0])])
+            >>> print(dev.state)
+            [0.+0.j 1.+0.j]
+            """
+            state = np.zeros(2**self.num_wires, dtype=self.C_DTYPE)
+            state = self._asarray(state, dtype=self.C_DTYPE)
+            self._qubit_state.getState(state)
+            return state
 
         @property
         def state_vector(self):
-            """Returns a handle to a StateVector object matching ``use_csingle`` precision."""
-            ket = np.ravel(self._state)
-            return StateVectorC64(ket) if self.use_csingle else StateVectorC128(ket)
+            """Returns a handle to the statevector."""
+            return self._qubit_state
 
         def _apply_state_vector(self, state, device_wires):
             """Initialize the internal state vector in a specified state.
@@ -328,6 +329,12 @@ if LQ_CPP_BINARY_AVAILABLE:
                     or broadcasted state of shape ``(batch_size, 2**len(wires))``
                 device_wires (Wires): wires that get initialized in the state
             """
+
+            if isinstance(state, self._qubit_state.__class__):
+                state_data = allocate_aligned_array(state.size, np.dtype(self.C_DTYPE), True)
+                self._qubit_state.getState(state_data)
+                state = state_data
+
             ravelled_indices, state = self._preprocess_state_vector(state, device_wires)
 
             # translate to wire labels used by device
@@ -336,12 +343,11 @@ if LQ_CPP_BINARY_AVAILABLE:
 
             if len(device_wires) == self.num_wires and Wires(sorted(device_wires)) == device_wires:
                 # Initialize the entire device state with the input state
-                self._state = self._reshape(state, output_shape)
+                state = self._reshape(state, output_shape).ravel(order="C")
+                self._qubit_state.UpdateData(state)
                 return
 
-            state = self._scatter(ravelled_indices, state, [2**self.num_wires])
-            state = self._reshape(state, output_shape)
-            self._state = self._asarray(state, dtype=self.C_DTYPE)
+            self._qubit_state.setStateVector(ravelled_indices, state)  # this operation on device
 
         def _apply_basis_state(self, state, wires):
             """Initialize the state vector in a specified computational basis state.
@@ -355,22 +361,23 @@ if LQ_CPP_BINARY_AVAILABLE:
             Note: This function does not support broadcasted inputs yet.
             """
             num = self._get_basis_state_index(state, wires)
-            self._state = self._create_basis_state(num)
+            self._create_basis_state(num)
 
-        def _apply_lightning_controlled(self, sim, operation):
+        def _apply_lightning_controlled(self, operation):
             """Apply an arbitrary controlled operation to the state tensor.
 
             Args:
-                sim (StateVectorC64, StateVectorC128): a state vector simulator
                 operation (~pennylane.operation.Operation): operation to apply
 
             Returns:
                 array[complex]: the output state tensor
             """
+            state = self.state_vector
+
             basename = "PauliX" if operation.name == "MultiControlledX" else operation.base.name
             if basename == "Identity":
                 return
-            method = getattr(sim, f"{basename}", None)
+            method = getattr(state, f"{basename}", None)
             control_wires = self.wires.indices(operation.control_wires)
             control_values = (
                 [bool(int(i)) for i in operation.hyperparameters["control_values"]]
@@ -386,7 +393,7 @@ if LQ_CPP_BINARY_AVAILABLE:
                 param = operation.parameters
                 method(control_wires, control_values, target_wires, inv, param)
             else:  # apply gate as an n-controlled matrix
-                method = getattr(sim, "applyControlledMatrix")
+                method = getattr(state, "applyControlledMatrix")
                 target_wires = self.wires.indices(operation.target_wires)
                 try:
                     method(
@@ -402,27 +409,24 @@ if LQ_CPP_BINARY_AVAILABLE:
                         operation.base.matrix, control_wires, control_values, target_wires, False
                     )
 
-        def apply_lightning(self, state, operations):
+        def apply_lightning(self, operations):
             """Apply a list of operations to the state tensor.
 
             Args:
-                state (array[complex]): the input state tensor
                 operations (list[~pennylane.operation.Operation]): operations to apply
 
             Returns:
                 array[complex]: the output state tensor
             """
-            state_vector = np.ravel(state)
-            sim = (
-                StateVectorC64(state_vector) if self.use_csingle else StateVectorC128(state_vector)
-            )
+            state = self.state_vector
 
             # Skip over identity operations instead of performing
             # matrix multiplication with it.
             for operation in operations:
-                if operation.name == "Identity":
+                name = operation.name
+                if name == "Identity":
                     continue
-                method = getattr(sim, operation.name, None)
+                method = getattr(state, name, None)
                 wires = self.wires.indices(operation.wires)
 
                 if method is not None:  # apply specialized gate
@@ -430,22 +434,20 @@ if LQ_CPP_BINARY_AVAILABLE:
                     param = operation.parameters
                     method(wires, inv, param)
                 elif (
-                    operation.name[0:2] == "C("
-                    or operation.name == "ControlledQubitUnitary"
-                    or operation.name == "MultiControlledX"
+                    name[0:2] == "C("
+                    or name == "ControlledQubitUnitary"
+                    or name == "MultiControlledX"
                 ):  # apply n-controlled gate
-                    self._apply_lightning_controlled(sim, operation)
+                    self._apply_lightning_controlled(operation)
                 else:  # apply gate as a matrix
                     # Inverse can be set to False since qml.matrix(operation) is already in
                     # inverted form
-                    method = getattr(sim, "applyMatrix")
+                    method = getattr(state, "applyMatrix")
                     try:
                         method(qml.matrix(operation), wires, False)
                     except AttributeError:  # pragma: no cover
                         # To support older versions of PL
                         method(operation.matrix, wires, False)
-
-            return np.reshape(state_vector, state.shape)
 
         # pylint: disable=unused-argument
         def apply(self, operations, rotations=None, **kwargs):
@@ -468,15 +470,7 @@ if LQ_CPP_BINARY_AVAILABLE:
                         f"Operations have already been applied on a {self.short_name} device."
                     )
 
-            if operations:
-                self._pre_rotated_state = self.apply_lightning(self._state, operations)
-            else:
-                self._pre_rotated_state = self._state
-
-            if rotations:
-                self._state = self.apply_lightning(np.copy(self._pre_rotated_state), rotations)
-            else:
-                self._state = self._pre_rotated_state
+            self.apply_lightning(operations)
 
         # pylint: disable=protected-access
         def expval(self, observable, shot_range=None, bin_size=None):
@@ -494,9 +488,11 @@ if LQ_CPP_BINARY_AVAILABLE:
                 Expectation value of the observable
             """
             if observable.name in [
-                "Identity",
                 "Projector",
             ]:
+                if self.shots is None:
+                    qs = qml.tape.QuantumScript([], [qml.expval(observable)])
+                    self.apply(self._get_diagonalizing_gates(qs))
                 return super().expval(observable, shot_range=shot_range, bin_size=bin_size)
 
             if self.shots is not None:
@@ -505,14 +501,10 @@ if LQ_CPP_BINARY_AVAILABLE:
                 samples = self.sample(observable, shot_range=shot_range, bin_size=bin_size)
                 return np.squeeze(np.mean(samples, axis=0))
 
-            # Initialization of state
-            ket = np.ravel(self._pre_rotated_state)
-
-            state_vector = StateVectorC64(ket) if self.use_csingle else StateVectorC128(ket)
             measurements = (
-                MeasurementsC64(state_vector)
+                MeasurementsC64(self.state_vector)
                 if self.use_csingle
-                else MeasurementsC128(state_vector)
+                else MeasurementsC128(self.state_vector)
             )
             if observable.name == "SparseHamiltonian":
                 csr_hamiltonian = observable.sparse_matrix(wire_order=self.wires).tocsr(copy=False)
@@ -552,9 +544,11 @@ if LQ_CPP_BINARY_AVAILABLE:
                 Variance of the observable
             """
             if observable.name in [
-                "Identity",
                 "Projector",
             ]:
+                if self.shots is None:
+                    qs = qml.tape.QuantumScript([], [qml.var(observable)])
+                    self.apply(self._get_diagonalizing_gates(qs))
                 return super().var(observable, shot_range=shot_range, bin_size=bin_size)
 
             if self.shots is not None:
@@ -563,14 +557,10 @@ if LQ_CPP_BINARY_AVAILABLE:
                 samples = self.sample(observable, shot_range=shot_range, bin_size=bin_size)
                 return np.squeeze(np.var(samples, axis=0))
 
-            # Initialization of state
-            ket = np.ravel(self._pre_rotated_state)
-
-            state_vector = StateVectorC64(ket) if self.use_csingle else StateVectorC128(ket)
             measurements = (
-                MeasurementsC64(state_vector)
+                MeasurementsC64(self.state_vector)
                 if self.use_csingle
-                else MeasurementsC128(state_vector)
+                else MeasurementsC128(self.state_vector)
             )
 
             if observable.name == "SparseHamiltonian":
@@ -603,14 +593,10 @@ if LQ_CPP_BINARY_AVAILABLE:
                 array[int]: array of samples in binary representation with shape
                     ``(dev.shots, dev.num_wires)``
             """
-            # Initialization of state
-            ket = np.ravel(self._state)
-
-            state_vector = StateVectorC64(ket) if self.use_csingle else StateVectorC128(ket)
             measurements = (
-                MeasurementsC64(state_vector)
+                MeasurementsC64(self.state_vector)
                 if self.use_csingle
-                else MeasurementsC128(state_vector)
+                else MeasurementsC128(self.state_vector)
             )
             if self._mcmc:
                 return measurements.generate_mcmc_samples(
@@ -630,12 +616,21 @@ if LQ_CPP_BINARY_AVAILABLE:
             Returns:
                 array[float]: list of the probabilities
             """
-            state_vector = self.state_vector
             return (
-                MeasurementsC64(state_vector)
+                MeasurementsC64(self.state_vector)
                 if self.use_csingle
-                else MeasurementsC128(state_vector)
+                else MeasurementsC128(self.state_vector)
             ).probs(wires)
+
+        # pylint: disable=attribute-defined-outside-init
+        def sample(self, observable, shot_range=None, bin_size=None, counts=False):
+            """Return samples of an observable."""
+            if observable.name != "PauliZ":
+                self.apply_lightning(observable.diagonalizing_gates())
+                self._samples = self.generate_samples()
+            return super().sample(
+                observable, shot_range=shot_range, bin_size=bin_size, counts=counts
+            )
 
         @staticmethod
         def _check_adjdiff_supported_measurements(
@@ -700,14 +695,11 @@ if LQ_CPP_BINARY_AVAILABLE:
                         "The number of qubits of starting_state must be the same as "
                         "that of the device."
                     )
-                ket = self._asarray(starting_state, dtype=self.C_DTYPE)
-            else:
-                if not use_device_state:
-                    self.reset()
-                    self.apply(tape.operations)
-                ket = self._pre_rotated_state
-            ket = ket.reshape(-1)
-            return StateVectorC64(ket) if self.use_csingle else StateVectorC128(ket)
+                self._apply_state_vector(starting_state, self.wires)
+            elif not use_device_state:
+                self.reset()
+                self.apply(tape.operations)
+            return self.state_vector
 
         def adjoint_jacobian(self, tape, starting_state=None, use_device_state=False):
             """Computes and returns the Jacobian with the adjoint method."""